% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_fun.R
\name{admm_arg_ctrl}
\alias{admm_arg_ctrl}
\title{Control ADMM Arguments}
\usage{
admm_arg_ctrl(
  max_iter = 100,
  rho = 1,
  kappa = 1.8,
  abs_tol = 1e-04,
  rel_tol = 0.01,
  eps = 1e-08,
  inner_verbose = F
)
}
\arguments{
\item{max_iter}{a positive integer, the number of loops in ADMM algorithm. Default is \code{100}.}

\item{rho}{a positive double scalar, a penalized parameter in the ADMM loop. Default is \code{1}.}

\item{kappa}{a postive double scalar, a relaxation pameter in the ADMM loop for the better convergnece. Recommended value is between 1.6 and 1.8(Default).}

\item{abs_tol}{a positive double scale, an absolute tolerance for stopping criterion in the ADMM loop.}

\item{rel_tol}{positive double scale, an relative tolerance for stopping criterion in the ADMM loop.}

\item{eps}{a positive scalar, a tolerance of estimates of parameter for the various purpose. Default is \code{1e-8}}

\item{inner_verbose}{blah}
}
\value{
a list object with arguments for customizing ADMM algorithm.
}
\description{
Control ADMM Arguments
}
