% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alarm.R
\docType{data}
\name{alarm}
\alias{alarm}
\title{The ALARM network}
\format{
THE ALARM dataset consists of 8 diagnoses, 16 finding and 13 intermediate variables (total 37 variables). Check \code{bnlearn:alarm} for more details.
}
\usage{
data(alarm)
}
\description{
ALARM network data from the
\href{https://www.bnlearn.com/bnrepository/discrete-medium.html#alarm}{Bayesian network repository}. The ALARM(A Logical Alarm Reduction Mechanism) is designed to give an alarm message system for patient monitoring. All the variables are multi-level (2, 3, or 4 levels) variable and there are 20,000 observations in the original data.
}
\references{
[1] Beinlich IA, Suermondt HJ, Chavez RM, Cooper GF (1989). “The ALARM monitoring system: A case study with two probabilistic inference techniques for belief networks.” In AIME 89, pp. 247–256. Springer.
}
\keyword{datasets}
