% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alarm_true_str.R
\docType{data}
\name{alarm_true_str}
\alias{alarm_true_str}
\title{The ALARM network true structure information}
\format{
A list object including:
\describe{
\item{A_true}{adjacency matrix of the true network}
\item{graph_true}{igraph object based on the adjacency matrix}
\item{n_edges}{the number of true edges}
}
}
\usage{
data(alarm_true_str)
}
\description{
True network information of ALARM network data from the
\href{https://www.bnlearn.com/bnrepository/discrete-medium.html#alarm}{Bayesian network repository}.
}
\references{
[1] Beinlich IA, Suermondt HJ, Chavez RM, Cooper GF (1989). “The ALARM monitoring system: A case study with two probabilistic inference techniques for belief networks.” In AIME 89, pp. 247–256. Springer.
}
\keyword{datasets}
