% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_fun.R
\name{dag_plot}
\alias{dag_plot}
\title{Draw a DAG plot}
\usage{
dag_plot(
  graph_true,
  data_type,
  font_size = 1,
  main,
  cex.main = 2,
  default_nodes_name = F,
  ...
)
}
\arguments{
\item{graph_true}{blah}

\item{data_type}{blah}

\item{font_size}{blah}

\item{main}{blah}

\item{cex.main}{blah}

\item{default_nodes_name}{blah}

\item{...}{blah}
}
\value{
blah
}
\description{
Draw a DAG plot
}
