% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_adj_mat.R
\name{gen_adj_mat}
\alias{gen_adj_mat}
\title{Generate a true adjacency matrix}
\usage{
gen_adj_mat(
  n_nodes,
  graph_type = c("bi", "rand", "sf", "sw", "tree"),
  seed = NA
)
}
\arguments{
\item{n_nodes}{positive integer scalar (larger than 1), the number of nodes of the desired graph.}

\item{graph_type}{character scalar, one of types of graphs in the simulation studies. See details below for more information.}

\item{seed}{integer scalar, in order to reproduce the same graph.}
}
\value{
\item{A_true}{a corresponding adjacecny matrix of the \code{graph_true}}
}
\description{
Generate a desired graph (with \code{igraph} format) with corresponding adjacency matrix as an binary matrix. If \eqn{(i,j)} is one, it implies there exists an edge \eqn{i -> j}.
}
\details{
There are seven graph types in the simulation studies: "bi" for Bipartite graph, "sf" for Scale-Free graph, "star" for Star-Out graph, "sw" for Small-World, "tree" for the tree with two childeren, "rand" for random DAG, and "chain" for a Chain graph. Except "rand" and "chain", all the other graphs generated based on R package \strong{igraph}. The Random DAG is generaated based on R package \strong{pcalg}.
}
