% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Generate a dataset using a true graph and corresponding adjacency matrix}
\usage{
gen_data(n_obs, A_true, W_true, seed = 1)
}
\arguments{
\item{n_obs}{positive interge scalar, the desired number of observations.}

\item{A_true}{binary matrix, a true adjacecny matrix from a desired graph.}

\item{W_true}{double matrix, a true parameter matrix.}

\item{seed}{integer for seed control, default is 1.}
}
\value{
a positive integer or double matrix, a dataset with the desired number of observations.
}
\description{
Generate a dataset using a true graph and corresponding adjacency matrix
}
