% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_fun.R
\name{gen_node_levels}
\alias{gen_node_levels}
\title{Random level number generate by node}
\usage{
gen_node_levels(types_by_node, max_n_levels = 2, seed = 1)
}
\arguments{
\item{types_by_node}{blah}

\item{max_n_levels}{blah}

\item{seed}{blah}
}
\value{
blah
}
\description{
Random level number generate by node
}
