% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_para.R
\name{gen_para}
\alias{gen_para}
\title{Generate true parameters using an adjacency matrix}
\usage{
gen_para(
  A_true,
  types_by_node = NULL,
  n_resps_by_node = NULL,
  intcpt = c("always", "for_child_only", "none"),
  range = c(0.5, 1),
  prob_sparse = 0,
  seed = 1
)
}
\arguments{
\item{A_true}{binary matrix, a true adjacecny matrix from a desired graph.}

\item{types_by_node}{blah}

\item{n_resps_by_node}{blah}

\item{intcpt}{blah}

\item{range}{a possible range of (absolute value of) parameters.}

\item{prob_sparse}{double scalar, define a sparseness of parameters within each parameter sub-matrix with \eqn{0\le prob_sparse \le 1} - default is 0. If \eqn{prob_sparse = 0}, all the parameter for each submatrix will be non-zeros. Not recommnedable, but, if \eqn{prob_sparse = 1}, all the parameter for each submatrix will be zeros. See details below for the parameter sub-matrix.}

\item{seed}{blah}
}
\value{
double matrix, true parameter matrix, \code{W_true} is returend with \code{n_levels} or \code{n_group_elmts} as an attribute.
}
\description{
Generate true parameters using an adjacency matrix
}
\details{
The parameter matrix can be divided into submatrices. For example, a submatrix (i,j) consists of parameters related to (node i) -> (node j). The \code{prob_sparse} controls the sparseness of parameters within each submatrix.
}
