% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_fun.R
\name{print.noteargis}
\alias{print.noteargis}
\title{Print a \code{noteargis} object}
\usage{
\method{print}{noteargis}(x, ...)
}
\arguments{
\item{x}{fitted \code{noteargis} object}

\item{\dots}{additional arguments if available}
}
\value{
Display \code{graph_est_by_lam} and \code{A_est_by_lam}
}
\description{
Print a \code{noteargis} object
}
