% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noteargis.R
\name{noteargis}
\alias{noteargis}
\title{Title}
\usage{
noteargis(
  data_input,
  data_type = c("c", "m"),
  n_resps_by_node = NULL,
  intcpt = NULL,
  lambdas = NULL,
  n_lams = NULL,
  admm_args = admm_arg_ctrl(),
  add_stop_rule = TRUE,
  fac_grp_lasso = FALSE,
  verbose = FALSE,
  fit_hist = FALSE
)
}
\arguments{
\item{data_input}{blah}

\item{data_type}{blah}

\item{n_resps_by_node}{blah}

\item{intcpt}{blah}

\item{lambdas}{blah}

\item{n_lams}{blah}

\item{admm_args}{blah}

\item{add_stop_rule}{blah}

\item{fac_grp_lasso}{blah}

\item{verbose}{blah}

\item{fit_hist}{blah}
}
\value{
blah blah
}
\description{
Title
}
